requirements.txt
langchain 
python-dotenv
ipykernel
langchain_groq
langchain_community
langchain_huggingface
streamlit
python-telegram-bot


